Выполнена модуляция
